Move MiniProtocolIx type and make it a newtype
It is used in array indexes, so useful to have it as a proper newtype to
avoid silly indexing mistakes.

+++ network-mux/src/Network/Mux.hs
@@ -171,7 +171,7 @@ muxStart tracer (MuxApplication ptcls) bearer = do
                  ])
       where
         minpix = 0
-        maxpix = length ptcls' - 1
+        maxpix = fromIntegral (length ptcls' - 1)
 
         codes   = [ miniProtocolNum ptcl | (ptcl, _, _) <- ptcls' ]
         mincode = minimum codes
